Reorganize browserify caching code
