+++ build-system/common/logging.js
@@ -19,6 +19,14 @@ const log = require('fancy-log');
 const {bold, yellow} = require('ansi-colors');
 const {isCiBuild} = require('./ci');
 
+/**
+ * Used by tests to wrap progres dots.
+ */
+const dotWrappingWidth = 150;
+
+/**
+ * Used by CI job scripts to print a prefix before top-level logging lines.
+ */
 let loggingPrefix = '';
 
 /**
@@ -89,6 +97,7 @@ function logWithoutTimestampLocalDev(...messages) {
 }
 
 module.exports = {
+  dotWrappingWidth,
   getLoggingPrefix,
   log,
   logLocalDev,
