+++ build-system/tasks/e2e/mocha-dots-reporter.js
@@ -25,8 +25,11 @@ const {
   EVENT_TEST_PENDING,
 } = Mocha.Runner.constants;
 const {Base} = Mocha.reporters;
+const {
+  icon,
+  nbDotsPerLine,
+} = require('../../test-configs/karma.conf').superDotsReporter;
 const {green, red, yellow} = require('ansi-colors');
-const {icon, nbDotsPerLine} = require('../karma.conf').superDotsReporter;
 const {reportTestFinished} = require('../report-test-status');
 
 /**
